% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{additional_european_pollen}
\alias{additional_european_pollen}
\title{\code{additional_european_pollen} dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 21 columns.
}
\usage{
additional_european_pollen
}
\description{
Additional European pollen samples. Part of the \link{SMPDSv2}.
}
\keyword{datasets}
