% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMSS}
\alias{AMSS}
\title{\code{AMSS} dataset}
\format{
A data frame (\code{tibble} object) with 38 rows and 23 variables:
\describe{
\item{source}{\code{character} Source of the dataset}
\item{site_name}{\code{character} Site name}
\item{entity_name}{\code{character} Entity name}
\item{latitude}{\code{numeric} Latitude}
\item{longitude}{\code{numeric} Longitude}
\item{elevation}{\code{numeric} Elevation}
\item{basin_size}{\code{character} Basin size}
\item{site_type}{\code{character} Site type}
\item{entity_type}{\code{character} Entity type}
\item{age_BP}{\code{character} Age BP}
\item{publication}{\code{character} Publication}
\item{doi}{\code{character} DOI}
\item{ID_BIOME}{\code{numeric} Biome identifier}
\item{ID_SAMPLE}{\code{numeric} Sample identifier}
\item{mi}{\code{numeric} Moisture index}
\item{gdd0}{\code{numeric} Growing degree days with 0 deg Celsius}
\item{mat}{\code{numeric} Mean annual temperature}
\item{mtco}{\code{numeric} Mean temperature of the coldest month}
\item{mtwa}{\code{numeric} Mean temperature of the warmest month}
\item{map}{\code{numeric} Mean annual precipitation}
\item{clean}{\code{data.frame} Clean pollen counts}
\item{intermediate}{\code{data.frame} Intermediate (grouped) pollen counts}
\item{amalgamated}{\code{data.frame} Amalgamated pollen counts}
}
}
\usage{
AMSS
}
\description{
The African Modern Surface Samples Dataset. Part of the \link{SMPDSv2}.
}
\references{
\insertRef{jolly1996representation}{smpds}

\insertRef{lebamba2009modern}{smpds}

\insertRef{julier2018modern}{smpds}

\insertRef{julier2019variability}{smpds}
}
\keyword{datasets}
