% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APD}
\alias{APD}
\title{\code{APD} dataset}
\format{
A data frame (\code{tibble} object) with 241622 rows and 7 variables:
\describe{
\item{...}{int ...}
}
}
\source{
Davis, B.A., Chevalier, M., Sommer, P., Carter, V.A., Finsinger, W., Mauri, A., Phelps, L.N., Zanon, M., Abegglen, R., Åkesson, C.M. and Alba-Sánchez, F., 2020. The Eurasian Modern Pollen Database (EMPD), version 2. Earth system science data, 12(4), pp.2423-2445. https://doi.org/10.5194/essd-12-2423-2020
}
\usage{
APD
}
\description{
The African Pollen Dataset.
}
\keyword{datasets}
