% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{australia_pollen}
\alias{australia_pollen}
\title{\code{australia_pollen} dataset}
\format{
A data frame (\code{tibble} object) with 1541 rows and 17 variables:
\describe{
\item{...}{int ...}
}
}
\usage{
australia_pollen
}
\description{
The Australian Pollen Dataset.
}
\keyword{datasets}
