% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate}
\alias{climate}
\title{\code{climate}}
\format{
An object of class \code{climate} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 24649 rows and 9 columns.
}
\usage{
climate
}
\description{
\code{climate}
}
\keyword{datasets}
