% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CMPD}
\alias{CMPD}
\title{\code{CMPD} dataset}
\format{
A data frame (\code{tibble} object) with 4208 rows and 22 variables:
\describe{
\item{...}{int ...}
}
}
\usage{
CMPD
}
\description{
The Chinese Modern Pollen Dataset by Ni Jian.
}
\keyword{datasets}
