% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRU_coords}
\alias{CRU_coords}
\title{\code{CRU_coords} dataset}
\format{
A data frame (\code{tibble} object) with 67420 rows and 3 variables:
\describe{
\item{latitude}{Numeric values for latitude in decimal degrees.}
\item{longitude}{Numeric values for longitude in decimal degrees.}
\item{elevation}{Numeric values for elevation in metres above sea level.}
}
}
\source{
https://crudata.uea.ac.uk/~timm/grid/CRU_TS_2_1.html
}
\usage{
CRU_coords
}
\description{
The CRU TS 2.1 dataset coordinates at 0.5 degrees resolution.
}
\references{
Mitchell, T.D. and Jones, P.D., 2005. An improved method of constructing a database of monthly climate observations and associated high‐resolution grids. International Journal of Climatology: A Journal of the Royal Meteorological Society, 25(6), pp.693-712. https://doi.org/10.1002/joc.1181
}
\keyword{datasets}
