% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{cru_mask}
\alias{cru_mask}
\title{Create land-sea mask for the CRU TS dataset}
\usage{
cru_mask(res = 0.5, coordinates = smpds::CRU_coords)
}
\arguments{
\item{res}{Numeric value for the mask resolution. Default: 0.5 degrees.}

\item{coordinates}{Reference data set with columns for \code{latitude},
\code{longitude} and \code{elevation}. Default:
\code{\link{CRU_coords}}.}
}
\value{
Table with land-sea mask:
\itemize{
\item \code{land = TRUE}: Grid cell with data provided by the CRU TS.
\item \code{land = FALSE}: Grid cell where data is not provided by the
CRU TS.
}
}
\description{
Create land-sea mask for the CRU TS dataset
}
\keyword{internal}
