% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dugerdil_pollen}
\alias{dugerdil_pollen}
\title{\code{dugerdil_pollen} dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 23 columns.
}
\usage{
dugerdil_pollen
}
\description{
Pollen samples provided by L. Dugerdil. Part of the \link{SMPDSv2}.
}
\references{
\insertRef{dugerdil2021climate}{smpds}
}
\keyword{datasets}
