% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMBSeCBIO}
\alias{EMBSeCBIO}
\title{\code{EMBSeCBIO} dataset}
\format{
A data frame (\code{tibble} object) with 149 rows and 23 variables:
\describe{
\item{...}{int ...}
}
}
\source{
10.17864/1947.309
}
\usage{
EMBSeCBIO
}
\description{
The Eastern Mediterranean-Black Sea-Caspian-Corridor region and biome
reconstructions database.
}
\references{
\insertRef{harrison2017embsecbio}{smpds}

\insertRef{harrison2021embsecbio}{smpds}
}
\keyword{datasets}
