% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMPDv2}
\alias{EMPDv2}
\title{\code{EMPDv2} dataset}
\format{
A data frame (\code{tibble} object) with 3508 rows and 17 variables:
\describe{
\item{...}{int ...}
}
}
\source{
https://doi.pangaea.de/10.1594/PANGAEA.909130
}
\usage{
EMPDv2
}
\description{
The Eurasian Modern Pollen Database (EMPD), version 2.
}
\references{
\insertRef{davis2020eurasian}{smpds}
}
\keyword{datasets}
