% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entity}
\alias{entity}
\title{\code{entity}}
\format{
An object of class \code{entity} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 24649 rows and 15 columns.
}
\usage{
entity
}
\description{
\code{entity}
}
\keyword{datasets}
