% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-biome.R
\name{extract_biome}
\alias{extract_biome}
\title{Extract biome}
\usage{
extract_biome(data, reference = smpds::PNV, buffer = 12000)
}
\arguments{
\item{data}{Table containing a geometry column.}

\item{reference}{Reference map with biomes, default: \code{smpds::PNV}.}

\item{buffer}{numeric. The radius of a buffer around each point from which to extract cell values. If the distance between the sampling point and the center of a cell is less than or equal to the buffer, the cell is included. The buffer can be specified as a single value, or as a vector of the length of the number of points.  If the data are not projected (latitude/longitude), the unit should be meters. Otherwise it should be in map-units (typically also meters).}
}
\value{
Tibble with matched biomes: dominant, sub-dominant, etc.
}
\description{
Extracts biome for a point (latitude, longitude) based on the map created by
Hengl et al., 2018.
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
data <- tibble::tibble(entity_name = "University of Reading",
                        latitude = 51.4414,
                        longitude = 0.9418) \%>\%
  sf::st_as_sf(x = ., coords = c("longitude", "latitude"))
data \%>\%
  extract_biome()
}
\seealso{
\code{\link[raster]{getValues}, \link[raster]{getValuesFocal}}
}
