% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaillard_pollen}
\alias{gaillard_pollen}
\title{\code{gaillard_pollen} dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 124 rows and 23 columns.
}
\usage{
gaillard_pollen
}
\description{
Pollen samples provided by M.J. Gaillard. Part of the \link{SMPDSv2}.
}
\references{
\insertRef{gaillard1992modern}{smpds}
}
\keyword{datasets}
