% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{gdd}
\alias{gdd}
\alias{gdd.numeric}
\alias{gdd.tbl_df}
\title{Calculate GDD}
\usage{
gdd(.data, ...)

\method{gdd}{numeric}(.data, baseline = 0, ...)

\method{gdd}{tbl_df}(.data, baseline = 0, ...)
}
\arguments{
\item{.data}{Object with daily values of temperature. The object can be a:
\itemize{
\item Numeric vector
\item Data frame (\code{tibble} object) with a column called \code{tmp},
with values of daily temperature.
}}

\item{...}{Optional parameters. Including the \code{baseline} for the
calculation of the Growing Degree Days.}

\item{baseline}{Numeric value to be used as the baseline for the calculation
of the Growing Degree Days, default: \code{0}.}
}
\value{
If the input is a
\itemize{
\item Numeric vector, it returns a single value for GDD.
\item Data frame (\code{tibble} object), it returns the same data frame
with an additional column called \code{gddX} (where \code{X} represents
the value of the \code{baseline}; e.g., \code{gdd0} when
\code{baseline = 0}), containing values for GDD.
}
}
\description{
Calculate Growing Degree Days above a baseline  \code{baseline}, (GDD) from
daily values of temperature.
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
data <- tibble::tibble(entity_name = "University of Reading",
                       latitude = 51.44140,
                       longitude = -0.9418,
                       elevation = 61)
data \%>\%
  smpds::gwr(varid = "tmp",
             reference = "inst/extdata/cru_ts4.04-clim-1961-1990-daily_tmp_1-5.nc") \%>\%
  smpds::pivot_data(varname = "tmp") \%>\%
  smpds::gdd()
}
\seealso{
\code{\link{plot_gdd}}

Other utils climate: 
\code{\link{mtco}()},
\code{\link{mtwa}()}
}
\concept{utils climate}
