% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Herzschuh}
\alias{Herzschuh}
\title{\code{Herzschuh} dataset}
\format{
A data frame (\code{tibble} object) with 2598 rows and 221 variables:
\describe{
\item{...}{int ...}
}
}
\source{
Herzschuh, U., Cao, X., Laepple, T., Dallmeyer, A., Telford, R.J., Ni, J., Chen, F., Kong, Z., Liu, G., Liu, K.B. and Liu, X., 2019. Position and orientation of the westerly jet determined Holocene rainfall patterns in China. Nature communications, 10(1), pp.1-8. https://doi.org/10.1038/s41467-019-09866-8
}
\usage{
Herzschuh
}
\description{
\code{Herzschuh} dataset
}
\keyword{datasets}
