% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{japanese_pollen}
\alias{japanese_pollen}
\title{\code{japanese_pollen} dataset}
\format{
A data frame (\code{tibble} object) with 94 rows and 23 variables:
\describe{
\item{...}{int ...}
}
}
\usage{
japanese_pollen
}
\description{
The Japanese Pollen Dataset.
}
\keyword{datasets}
