% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{mask_nc}
\alias{mask_nc}
\title{Mask NetCDF}
\usage{
mask_nc(.data, mask = cru_mask())
}
\arguments{
\item{.data}{3D matrix with \code{latitude}, \code{longitude} and
\code{time}.}

\item{mask}{2D matrix with all combinations of \code{latitude},
\code{longitude} and a third variable called \code{land} with logical
values to indicated whether a grid cell should be used or ignored.
\itemize{
\item{\code{land = TRUE} }{ use this value.}
\item{\code{land = FALSE} }{ ignore this value.}
}}
}
\value{
2D version of \code{.data}, including \code{latitute},
\code{longitude}, and variables with the format \code{T#}, for each
time step in \code{.data}: \code{T1}, \code{T2}, ..., \code{Tk}.
}
\description{
Mask NetCDF variable.
}
\keyword{internal}
