% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{mat}
\alias{mat}
\alias{mat.numeric}
\alias{mat.tbl_df}
\title{Calculate MAT}
\usage{
mat(.data, ...)

\method{mat}{numeric}(.data, pb = NULL, ...)

\method{mat}{tbl_df}(.data, cpus = 1, ...)
}
\arguments{
\item{.data}{Object with daily values of temperature. The object can be a:
\itemize{
\item Numeric vector
\item Data frame (\code{tibble} object) with a column called \code{tmp},
with values of daily temperature (a \code{list}).
}}

\item{...}{Optional parameters, including \code{cpus} and \code{pb}.}

\item{pb}{Function to signal the execution of one cycle (e.g.
\code{progressr::progressor(...)}). Default: \code{NULL}, do nothing.}

\item{cpus}{Numeric value with the number of CPUs to be used for the
computation. Default: \code{1}, serial computation.}
}
\value{
If the input (\code{.data}) is a
\itemize{
\item Numeric vector, it returns a single value for MAT.
\item Data frame (\code{tibble} object), it returns the same data frame
with an additional column called \code{mat}, containing values for MAT.
}
}
\description{
Calculate Mean Annual Temperature (MAT) from daily values of temperature.
}
\seealso{
\code{\link{plot_mat}}

Other utils climate: 
\code{\link{gdd}()},
\code{\link{mi}()},
\code{\link{mtco}()},
\code{\link{mtwa}()}
}
\concept{utils climate}
