% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{mi}
\alias{mi}
\alias{mi.tbl_df}
\title{Calculate MI}
\usage{
mi(.data, ...)

\method{mi}{tbl_df}(.data, cpus = 1, ...)
}
\arguments{
\item{.data}{Data frame (\code{tibble} object) (see the details section).}

\item{...}{Optional parameters, including \code{cpus}.}

\item{cpus}{Numeric value with the number of CPUs to be used for the
computation. Default: \code{1}, serial computation.}
}
\value{
Data frame (\code{tibble} object) with the same input data and the
addition of a column called \code{mi}, containing values for MI.
}
\description{
Calculate Moisture Index (MI) from daily values of precipitation, sunshine
fraction and temperature.
}
\details{
The input (\code{.data} object) data, should contain the following
columns:
\itemize{
\item \code{latitude}: numeric values for latitude (decimal degrees).
\item \code{elevation}: numeric values for elevation (m A.S.L.)
\item \code{pre}: list of numeric values for daily precipitation
(mm / day).
\item \code{sf}: list of numeric values for daily sunshine fraction (\%).
\item \code{tmp}: list of numeric values for daily temperature
(degrees Celsius).
}
}
\seealso{
\code{\link{plot_mi}}

Other utils climate: 
\code{\link{gdd}()},
\code{\link{mat}()},
\code{\link{mtco}()},
\code{\link{mtwa}()}
}
\concept{utils climate}
