% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moroccan_pollen}
\alias{moroccan_pollen}
\title{\code{moroccan_pollen} dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 23 columns.
}
\usage{
moroccan_pollen
}
\description{
Moroccan pollen samples. Part of the \link{SMPDSv2}.
}
\references{
\insertRef{hamouri1991changements}{smpds}

\insertRef{lamb1995vegetational}{smpds}

\insertRef{cheddadi1998holocene}{smpds}

\insertRef{el201418}{smpds}

\insertRef{cheddadi2015history}{smpds}

\insertRef{campbell2017environmental}{smpds}

\insertRef{cheddadi2017microrefugia}{smpds}

\insertRef{fletcher2017ams}{smpds}

\insertRef{zielhofer2017atlantic}{smpds}
}
\keyword{datasets}
