% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{mtco}
\alias{mtco}
\alias{mtco.numeric}
\alias{mtco.tbl_df}
\title{Calculate MTCO}
\usage{
mtco(.data, ...)

\method{mtco}{numeric}(.data, pb = NULL, ...)

\method{mtco}{tbl_df}(.data, cpus = 1, ...)
}
\arguments{
\item{.data}{Object with daily values of temperature. The object can be a:
\itemize{
\item Numeric vector
\item Data frame (\code{tibble} object) with a column called \code{tmp},
with values of daily temperature (a \code{list}).
}}

\item{...}{Optional parameters, including \code{cpus} and \code{pb}.}

\item{pb}{Function to signal the execution of one cycle (e.g.
\code{progressr::progressor(...)}). Default: \code{NULL}, do nothing.}

\item{cpus}{Numeric value with the number of CPUs to be used for the
computation. Default: \code{1}, serial computation.}
}
\value{
If the input (\code{.data}) is a
\itemize{
\item Numeric vector, it returns a single value for MTCO.
\item Data frame (\code{tibble} object), it returns the same data frame
with an additional column called \code{mtco}, containing values for MTCO.
}
}
\description{
Calculate Mean Temperature of the COldest month (MTCO) from daily values
of temperature.
}
\seealso{
\code{\link{plot_mtco}}

Other utils climate: 
\code{\link{gdd}()},
\code{\link{mat}()},
\code{\link{mi}()},
\code{\link{mtwa}()}
}
\concept{utils climate}
