% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{mtwa}
\alias{mtwa}
\alias{mtwa.numeric}
\alias{mtwa.tbl_df}
\title{Calculate MTWA}
\usage{
mtwa(.data, ...)

\method{mtwa}{numeric}(.data, ...)

\method{mtwa}{tbl_df}(.data, ...)
}
\arguments{
\item{.data}{Object with daily values of temperature. The object can be a:
\itemize{
\item Numeric vector
\item Data frame (\code{tibble} object) with a column called \code{tmp},
with values of daily temperature.
}}

\item{...}{Optional parameters (not used).}
}
\value{
If the input is a
\itemize{
\item Numeric vector, it returns a single value for MTWA.
\item Data frame (\code{tibble} object), it returns the same data frame
with an additional column called \code{mtco}, containing values for MTWA.
}
}
\description{
Calculate Mean Temperature of the WArmest month (MTWA) from daily values
of temperature.
}
