% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neotropics_pollen}
\alias{neotropics_pollen}
\title{\code{neotropics_pollen} dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 636 rows and 23 columns.
}
\usage{
neotropics_pollen
}
\description{
Neotropics pollen samples. Part of the \link{SMPDSv2}.
}
\references{
\insertRef{bush2021modern}{smpds}
}
\keyword{datasets}
