% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset.R
\name{normalise_taxa}
\alias{normalise_taxa}
\title{Normalise taxa/columns counts}
\usage{
normalise_taxa(.data, cols = 1, scale = 100)
}
\arguments{
\item{.data}{Data frame.}

\item{cols}{Numeric vector with the columns to be ignored (metadata columns).}

\item{scale}{Numeric scale factor, default = 100.}
}
\value{
Data frame with normalised taxa/columns counts.
}
\description{
Normalises the taxa/columns counts by entity/row. The \code{scale} parameter
can be used to scale the normalised values, defaults to 100.
}
