% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{north_america_pollen}
\alias{north_america_pollen}
\title{\code{north_america_pollen} dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6219 rows and 23 columns.
}
\usage{
north_america_pollen
}
\description{
North America  pollen samples. Part of the \link{SMPDSv2}.
}
\keyword{datasets}
