% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Phelps}
\alias{Phelps}
\title{\code{Phelps} dataset}
\format{
A data frame (\code{tibble} object) with 102 rows and 651 variables:
\describe{
\item{...}{int ...}
}
}
\source{
https://doi.pangaea.de/10.1594/PANGAEA.905309
}
\usage{
Phelps
}
\description{
Dataset provided by \insertCite{phelps2020asymmetric;textual}{smpds}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
