% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{pivot_data}
\alias{pivot_data}
\title{Pivot data}
\usage{
pivot_data(
  .data,
  timestep = "^T[0-9]*",
  cols = colnames(.data) \%>\% stringr::str_detect(timestep, negate = TRUE) \%>\%
    which(),
  scale = 1,
  add = 0,
  varname = "value"
)
}
\arguments{
\item{.data}{Data frame (\code{tibble} object) obtained with
\code{\link{gwr}}.}

\item{timestep}{String with the pattern used to name each data point at each
time-step. Default: \code{"^T[0-9]*"}.}

\item{cols}{Metadata columns (\code{latitude}, \code{longitude}, etc.), to
be excluded from the data pivoting. Default: excluded columns matching
the \code{timestep} pattern, for the default pattern, ignores columns
with the prefix \code{T} and digits as suffix (e.g., \code{T1},
\code{T2}, ..., \code{Tt}, etc.).}

\item{scale}{Numeric value to scale the data. Default: \code{1}, no scaling.}

\item{add}{Numeric value to be added/subtracted from the data points.
Default: \code{0}, don't add anything.}

\item{varname}{Output variable name. Default: \code{"value"}.}
}
\value{
Data frame (\code{tibble} object) with a new column named according
to the string passed with \code{value}, this new column contains a list
of the data points at each time step for each observation/row.
}
\description{
Pivot data obtained with the function \code{\link{gwr}}. Groups all the data
points for each entity/row into a single list of elements.
}
\examples{
\dontrun{
`\%>\%` <- magrittr::`\%>\%`
data <- tibble::tibble(entity_name = "University of Reading",
                       latitude = 51.44140,
                       longitude = -0.9418,
                       elevation = 61)
smpds::gwr(.ref = "/path/to/reference-tmp.nc",
           .tar = data,
           varid = "tmp") \%>\%
  smpds::pivot_data(varname = "tmp")
}
}
