% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PNV_1km}
\alias{PNV_1km}
\title{\code{PNV_1km} dataset}
\format{
An object of class \code{RasterBrick} of dimension 17924 x 43200 x 1.
}
\source{
\insertRef{hengl2018global1kmres}{smpds}
}
\usage{
PNV_1km
}
\description{
Global Maps of Potential Natural Vegetation at 1 km resolution, version 4.
}
\references{
\insertRef{hengl2018global}{smpds}
}
\keyword{datasets}
