% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-biome.R
\name{pnv_classes}
\alias{pnv_classes}
\title{PNV classes}
\usage{
pnv_classes()
}
\value{
Data frame (\code{tibble} object) with three columns:
\itemize{
\item \code{ID_BIOME}: an unique identification number for each biome.
\item \code{description}: string witha description of each biome.
\item \code{colour}: hexadecimal colour code used to represent each biome.
}
}
\description{
Potential Natural Vegetation (PNV) classes based on the map created by
\insertCite{hengl2018global1kmres;textual}{smpds}.
}
\references{
\insertAllCited{}
}
