% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollen_count}
\alias{pollen_count}
\title{\code{pollen_count}}
\format{
An object of class \code{pollen_count} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1589561 rows and 4 columns.
}
\usage{
pollen_count
}
\description{
\code{pollen_count}
}
\keyword{datasets}
