% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{process_apd}
\alias{process_apd}
\title{Process the African Pollen Database}
\usage{
process_apd(path, ext = "ascii", delim = ";", comment = "#", ...)
}
\arguments{
\item{path}{Path to where the raw data files are located.}

\item{ext}{Extension of the raw data files.}

\item{delim}{Single character used to separate fields within a record.}

\item{comment}{A string used to identify comments. Any text after the
comment characters will be silently ignored.}

\item{...}{Additional parametres to be used by \code{readr::read_delim}}
}
\value{
List with a tibble for each of the raw data files
}
\description{
This function loads all the raw data from the African Pollen Database into
a single object.
}
\keyword{internal}
