% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset.R
\name{rm_na_taxa}
\alias{rm_na_taxa}
\title{Remove missing taxa/columns}
\usage{
rm_na_taxa(.data, cols = 1)
}
\arguments{
\item{.data}{Data frame.}

\item{cols}{Numeric vector with the columns to be ignored (metadata columns).}
}
\value{
Reduced data frame with non-missing taxa/column.
}
\description{
Removes taxa/columns that are missing.
}
