% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset.R
\name{rm_zero_taxa}
\alias{rm_zero_taxa}
\title{Remove taxa/columns with zeros}
\usage{
rm_zero_taxa(.data, cols = 1)
}
\arguments{
\item{.data}{Data frame.}

\item{cols}{Numeric vector with the columns to be ignored (metadata columns).}
}
\value{
Reduced data frame without taxa/columns of zeros.
}
\description{
Removes taxa/columns that are equal to zero
}
