% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smpds-package.R
\docType{package}
\name{smpds-package}
\alias{smpds}
\alias{smpds-package}
\title{smpds: The SPECIAL Modern Pollen Data}
\description{
Provides functions to facilitate the access to the SPECIAL Modern Pollen DataSet (SMPDS) and to reconstruct climate variables, like Mean Temperature of the COldest month (MTCO), Mean Temperature of the WArmest month (MTWA) and Growing Degree Days (GDD). Also, it provides functions to find the Potential Natural Vegetation (PNV) of a given site.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/special-uor/smpds/}
  \item \url{https://special-uor.github.io/smpds/}
  \item \url{https://research.reading.ac.uk/palaeoclimate/}
  \item Report bugs at \url{https://github.com/special-uor/smpds/issues/}
}

}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{r.villegas-diaz@outlook.com} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

Authors:
\itemize{
  \item Sandy P. Harrison \email{s.p.harrison@reading.ac.uk} (\href{https://orcid.org/0000-0001-5687-1903}{ORCID})
}

Other contributors:
\itemize{
  \item SPECIAL Research Group @ University of Reading [copyright holder]
}

}
\keyword{internal}
