% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMPDSv1}
\alias{SMPDSv1}
\title{\code{SMPDSv1} dataset}
\format{
A data frame (\code{tibble} object) with 6345 rows and 23 variables:
\describe{
\item{...}{int ...}
}
}
\source{
\insertRef{rdgdr194}{smpds}
}
\usage{
SMPDSv1
}
\description{
SPECIAL Modern Pollen Data for climate reconstructionS, version 1.
}
\keyword{datasets}
