% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMPDSv2}
\alias{SMPDSv2}
\title{\code{SMPDSv2} dataset}
\format{
A data frame (\code{tibble} object) with 24652 rows and 24 variables:
\describe{
\item{...}{int ...}
}
}
\usage{
SMPDSv2
}
\description{
SPECIAL Modern Pollen Data for climate reconstructionS, version 2.
}
\references{
\insertRef{davis2020eurasian}{smpds}

\insertRef{finsinger2007modern}{smpds}

\insertRef{rdgdr194}{smpds}

\insertRef{rdgdr294}{smpds}

\insertRef{herzschuh2019position}{smpds}

\insertRef{phelps2020asymmetric}{smpds}

\insertRef{vincens2007african}{smpds}

\insertRef{williams2018neotoma}{smpds}
}
\keyword{datasets}
