% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-db.R
\name{snapshot}
\alias{snapshot}
\alias{snapshot.MariaDBConnection}
\alias{snapshot.character}
\alias{snapshot.numeric}
\alias{snapshot.tbl_df}
\alias{snapshot.tbl}
\alias{snapshot.data.frame}
\alias{snapshot.default}
\title{Create a snapshot of the data linked to entities}
\usage{
snapshot(x, ...)

\method{snapshot}{MariaDBConnection}(x, ..., ID_ENTITY, ID_SITE, entity_name, site_name, quiet = TRUE)

\method{snapshot}{character}(x, ..., use_site_name = FALSE)

\method{snapshot}{numeric}(x, ..., use_id_site = FALSE)

\method{snapshot}{tbl_df}(x, ...)

\method{snapshot}{tbl}(x, ...)

\method{snapshot}{data.frame}(x, ...)

\method{snapshot}{default}(x, ...)
}
\arguments{
\item{x}{This object accepts different classes. If the given object is
a database connection, then extracts data from the database using the
\code{ID_SITE}, \code{ID_ENTITY} or \code{entity_name} (these should be provided after
the connection object). Alternatively, if the given object is a vector,
then it will retrieve the records from an internal snapshot of the
database, included in this package.}

\item{...}{Optional parameters.}

\item{ID_ENTITY}{Optional, if \code{ID_SITE}, \code{entity_name} or \code{site_name} are
provided.}

\item{ID_SITE}{Optional, if \code{ID_ENTITY}, \code{entity_name} or \code{site_name} are
provided.}

\item{entity_name}{Optional, if \code{ID_SITE}, \code{ID_ENTITY} or \code{site_name} are
provided.}

\item{site_name}{Optional, if \code{ID_SITE}, \code{ID_ENTITY} or \code{entity_name} are
provided.}

\item{quiet}{Boolean flag to indicate if queries should be displayed.}

\item{use_site_name}{Boolean flag to indicate whether to search using
\code{entity_name} (default) or \code{site_name}, using the values in \code{x}.}

\item{use_id_site}{Boolean flag to indicate whether to search using
\code{ID_ENTITY} (default) or \code{ID_SITE}, using the values in \code{x}.}
}
\value{
List with the individual tables.
}
\description{
Create a snapshot of the data linked to entities. Including metadata,
climate & vegetation reconstructions and pollen counts.
}
\examples{

\dontrun{
conn <- dabr::open_conn_mysql(dbname = "SMPDSv2",
                              password = rstudioapi::askForPassword())
# Using the entity name
snp1 <- smpds::snapshot(conn, entity_name = "juodonys_core")
snp1

# Using the site name
snp2 <- smpds::snapshot(conn, site_name = "Petresiunai")
snp2

# Using the ID_ENTITY
snp3 <- smpds::snapshot(conn, ID_ENTITY = 1)
snp3

# Using the ID_SITE
snp4 <- smpds::snapshot(conn, ID_SITE = 2)
snp4
}
# Using the entity name
snp1 <- smpds::snapshot("juodonys_core")
snp1

# Using the site name
snp2 <- smpds::snapshot("Petresiunai", use_site_name = TRUE)
snp2
# Using the ID_ENTITY
snp1 <- smpds::snapshot(1)
snp1

# Using the ID_SITE
snp2 <- smpds::snapshot(2, use_id_site = TRUE)
snp2
# Using the entity table
`\%>\%` <- magrittr::`\%>\%`
snp1 <- smpds::entity \%>\%
  dplyr::slice(1) \%>\%
  smpds::snapshot()
snp1

# Using a custom data frame (`tibble` object) with site names
snp2 <- tibble::tibble(
    site_name = c("Aligol lake", "Big Sandy Creek")
  ) \%>\%
  smpds::snapshot()
snp2
}
