% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset.R
\name{sort_taxa}
\alias{sort_taxa}
\title{Sort taxa/columns alphabetically}
\usage{
sort_taxa(.data, cols = 1)
}
\arguments{
\item{.data}{Data frame.}

\item{cols}{Numeric vector with the columns to be ignored (metadata columns).}
}
\value{
Data frame with taxa/column alphabetically sorted.
}
\description{
Sorts taxa/columns alphabetically, ignoring the columns in the positions
indicated by \code{col}.
}
