% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{south_america_pollen}
\alias{south_america_pollen}
\title{\code{south_america_pollen} dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 424 rows and 24 columns.
}
\usage{
south_america_pollen
}
\description{
South America  pollen samples. Part of the \link{SMPDSv2}.
}
\keyword{datasets}
