% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{southern_hemisphere_pollen}
\alias{southern_hemisphere_pollen}
\title{\code{southern_hemisphere_pollen} dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 76 rows and 24 columns.
}
\usage{
southern_hemisphere_pollen
}
\description{
Southern hemisphere  pollen samples. Part of the \link{SMPDSv2}.
}
\keyword{datasets}
