% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SSMPD}
\alias{SSMPD}
\title{\code{SSMPD} dataset}
\format{
A data frame (\code{tibble} object) with 124 rows and 183 variables:
\describe{
\item{...}{int ...}
}
}
\source{
https://doi.org/10.1007/BF00190697
}
\usage{
SSMPD
}
\description{
Dataset provided by \insertCite{gaillard1992modern;textual}{smpds}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
