% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{subset_coords}
\alias{subset_coords}
\title{Subset data}
\usage{
subset_coords(.data, latitude, longitude, buffer)
}
\arguments{
\item{.data}{2D matrix with columns called \code{latitude} and
\code{longitude}.}

\item{latitude}{Numeric value for reference \code{latitude}.}

\item{longitude}{Numeric value for reference \code{longitude}.}

\item{buffer}{Numeric value to be used as the boundary for the search area:
\itemize{
\item \code{latitude} < \code{.tar$latitude + buffer}
\item \code{latitude} > \code{.tar$latitude - buffer}
\item \code{longitude} < \code{.tar$longitude + buffer}
\item \code{longitude} > \code{.tar$longitude - buffer}
}}
}
\value{
Filtered 2D matrix.
}
\description{
Subset data using coordinates (\code{latitute} and \code{longitude}) and
\code{buffer}.
}
\keyword{internal}
