% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tatiana_pollen}
\alias{tatiana_pollen}
\title{\code{tatiana_pollen} dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 144 rows and 23 columns.
}
\usage{
tatiana_pollen
}
\description{
Pollen samples provided by T. Blyakharchuk. Part of the \link{SMPDSv2}.
}
\keyword{datasets}
