% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxon_name}
\alias{taxon_name}
\title{\code{taxon_name}}
\format{
An object of class \code{taxon_name} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5333 rows and 2 columns.
}
\usage{
taxon_name
}
\description{
\code{taxon_name}
}
\keyword{datasets}
