% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dataset.R
\name{total_taxa}
\alias{total_taxa}
\title{Add the taxa/columns counts by entity/row}
\usage{
total_taxa(.data, cols = 1)
}
\arguments{
\item{.data}{Data frame.}

\item{cols}{Numeric vector with the columns to be ignored (metadata columns).}
}
\value{
Data frame including a new column, \code{total}, with taxa counts.
}
\description{
Adds the taxa/columns counts by entity/row and inserts the a new column,
\code{total}, after the last index in \code{cols}, the metadata columns.
}
