% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{tps}
\alias{tps}
\title{Thin plate spline regression}
\usage{
tps(
  .data,
  var,
  resolution = 0.5,
  land_borders = rnaturalearth::ne_countries(scale = "small", returnclass = "sf"),
  check_data = TRUE,
  z_var = NULL,
  z_mode = "independent",
  z_ref = NULL,
  cpus = 1,
  ...
)
}
\arguments{
\item{.data}{Data frame with spatial and climate data. The following are
expected:
\itemize{
\item \strong{Latitude}, named: \code{latitude}, \code{lat} or \code{y}.
\item \strong{Longitude}, named: \code{longitude}, \code{long}, \code{lon}
or \code{y}.
\item \strong{Main variable}, named: value of \code{var}.
}}

\item{var}{String with the name of the climate variable to interpolate.}

\item{resolution}{Numeric value with the resolution (degrees) to interpolate.}

\item{land_borders}{Data frame with polygons to represent land borders
(e.g. continents, countries, counties, etc.).
Default: \code{rnaturalearth::ne_countries}.}

\item{check_data}{Boolean flag to indicate whether \code{.data} should be checked
or not (i.e. validate coordinates and main variable).
Default: \code{TRUE}}

\item{z_var}{String with the name of the variable containing information for
elevation. If this is given, then the elevation is used for the
interpolation.}

\item{z_mode}{String with the mode in which the elevation should be used:
\itemize{
\item \code{z_mode = "independent"} (Default), use the elevation as
another independent variable to predict \code{var}.
\item \code{z_mode = "covariate"}, use the elevation as a linear
covariate to predict \code{var}.
}}

\item{z_ref}{Raster object or path to raster object with grid containing
elevation information. For example the ETOPO5 (Earth topography 5 arc
minute) data set.}

\item{cpus}{Numeric value with the number of CPUs to use in the computation
of the elevations for the interpolation grid.}

\item{...}{
  Arguments passed on to \code{\link[fields:Tps]{fields::Tps}}
  \describe{
    \item{\code{m}}{
A polynomial function of degree (m-1) will be  
included in the model as the drift (or spatial trend) component. 
Default is the value such that 2m-d is greater than zero where d is the 
dimension of x. 
}
    \item{\code{p}}{
Polynomial power for Wendland radial basis functions. Default is 2m-d 
where d is the dimension of x.  
}
    \item{\code{scale.type}}{
The independent variables and knots are scaled to the specified 
scale.type. 
By default the scale type is "range", whereby 
the locations are transformed  
 to the interval (0,1) by forming (x-min(x))/range(x) for each x. 
Scale type of "user" allows specification of an x.center and
x.scale by the 
user. The default for "user" is mean 0 and standard deviation 1. Scale 
type of "unscaled" does not scale the data.  
}
    \item{\code{method}}{ Determines what "smoothing" parameter should be
     used. The default is to estimate standard GCV Other choices are:
     GCV.model, GCV.one, RMSE, pure error and REML. The differences
     are explained in the Krig help file.}
    \item{\code{GCV}}{If TRUE the decompositions are done to efficiently evaluate
 the estimate, GCV function and likelihood at multiple values
 of lambda. }
    \item{\code{miles}}{If TRUE great circle distances are in miles if FALSE
distances are in kilometers}
  }}
}
\value{
\code{tibble} object with interpolated values.
}
\description{
Thin plate spline regression
}
