% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{zissou1_palette}
\alias{zissou1_palette}
\title{Zissou1 colour palette}
\usage{
zissou1_palette(n = 100)
}
\description{
This is a colour palette created with the following function:
\code{wesanderson::wes_palette("Zissou1", 100, type = "continuous")}
Wrapped as a standalone function to remove dependency of the \code{wesanderson}
package.
}
\keyword{internal}
