% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APD}
\alias{APD}
\title{\code{APD} dataset}
\format{
A data frame (\code{tibble} object) with 90 rows and 23 variables:
\describe{
\item{...}{int ...}
}
}
\source{
http://fpd.sedoo.fr/fpd/bibli.do
}
\usage{
APD
}
\description{
The African Pollen Dataset.
}
\references{
\insertRef{vincens2007african}{smpds}
}
\keyword{datasets}
