% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{APD}
\alias{APD}
\title{\code{APD} dataset}
\format{
A data frame (\code{tibble} object) with 72 rows and 255 variables:
\describe{
\item{...}{int ...}
}
}
\source{
http://fpd.sedoo.fr/fpd/bibli.do
}
\usage{
APD
}
\description{
The African Pollen Dataset.
}
\references{
Vincens A, Lézine A, Buchet G, Lewden D, Le Thomas A (2007).
\dQuote{African pollen database inventory of tree and shrub pollen types.}
\emph{Review of Palaeobotany and Palynology}, \bold{145}(1-2), 135--141.
\doi{10.1016/j.revpalbo.2006.09.004}.
}
\keyword{datasets}
