% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-biome.R
\name{biome_name}
\alias{biome_name}
\title{Biome names}
\usage{
biome_name(ID)
}
\arguments{
\item{ID}{Numeric value linked to a Biome in provided by Hengl et al., 2018.}
}
\value{
Table (\code{tibble} object) with Biome metadata.
}
\description{
Obtain Biome names from the Hengl et al., 2018, using the \code{ID_BIOME}.
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
data <- tibble::tibble(entity_name = "University of Reading",
                        latitude = 51.4414,
                        longitude = 0.9418) \%>\%
  sf::st_as_sf(x = ., coords = c("longitude", "latitude"))
data \%>\%
  extract_biome() \%>\%
  .$ID_BIOME \%>\%
  biome_name()
}
