% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-biome.R
\name{extract_biome}
\alias{extract_biome}
\alias{extract_biome.tbl_df}
\alias{extract_biome.sf}
\alias{parallel_extract_biome}
\title{Extract biome}
\usage{
extract_biome(.data, ...)

\method{extract_biome}{tbl_df}(.data, ..., reference = smpds::PNV, buffer = 12000, all = FALSE)

\method{extract_biome}{sf}(.data, ..., reference = smpds::PNV, buffer = 12000, all = FALSE)

parallel_extract_biome(
  .data,
  reference = smpds::PNV,
  buffer = 12000,
  cpus = 2,
  all = FALSE
)
}
\arguments{
\item{.data}{Table containing a geometry column.}

\item{...}{
  Arguments passed on to \code{\link[=extract_biome.tbl_df]{extract_biome.tbl_df}}
  \describe{
    \item{\code{}}{}
  }}

\item{reference}{Reference map with biomes, default: \code{smpds::PNV}.}

\item{buffer}{numeric. The radius of a buffer around each point from which to extract cell values. If the distance between the sampling point and the center of a cell is less than or equal to the buffer, the cell is included. The buffer can be specified as a single value, or as a vector of the length of the number of points.  If the data are not projected (latitude/longitude), the unit should be meters. Otherwise it should be in map-units (typically also meters).}

\item{all}{Boolean flag to indicate whether or not to return all the detected
biomes, default: \code{FALSE} (dominant biome only).}

\item{cpus}{Number of CPUs to be used in parallel, default = 2.}
}
\value{
Tibble with original data and matched biomes:
\itemize{
\item{if \code{all = FALSE} (default) }{ Only returns the dominant biome:
\code{ID_BIOME}}
\item{if \code{all = TRUE} }{ Returns all the detected biomes:
\code{ID_BIOME} and \code{px}, the number of pixels detected for each
biome.}
}
}
\description{
Extracts biome for a point (latitude, longitude) based on the map created by
Hengl et al., 2018.
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
data <- tibble::tibble(entity_name = "University of Reading",
                        latitude = 51.44140,
                        longitude = 0.9418)
data \%>\%
  extract_biome()
}
\seealso{
Other utils biome: 
\code{\link{biome_name}()},
\code{\link{plot_biome}()}
}
\concept{utils biome}
