% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{gdd}
\alias{gdd}
\alias{gdd.numeric}
\alias{gdd.tbl_df}
\title{Calculate GDD}
\usage{
gdd(.data, ...)

\method{gdd}{numeric}(.data, baseline = 0, pb = NULL, ...)

\method{gdd}{tbl_df}(.data, baseline = 0, cpus = 1, ...)
}
\arguments{
\item{.data}{Object with daily values of temperature. The object can be a:
\itemize{
\item Numeric vector
\item Data frame (\code{tibble} object) with a column called \code{tmp},
with values of daily temperature.
}}

\item{...}{Optional parameters, including the \code{baseline} for the
calculation of the Growing Degree Days, \code{cpus} and \code{pb}.}

\item{baseline}{Numeric value to be used as the baseline for the calculation
of the Growing Degree Days, default: \code{0}.}

\item{pb}{Function to signal the execution of one cycle (e.g.
\code{progressr::progressor(...)}). Default: \code{NULL}, do nothing.}

\item{cpus}{Numeric value with the number of CPUs to be used for the
computation. Default: \code{1}, serial computation.}
}
\value{
If the input (\code{.data}) is a
\itemize{
\item Numeric vector, it returns a single value for GDD.
\item Data frame (\code{tibble} object), it returns the same data frame
with an additional column called \code{gddX} (where \code{X} represents
the value of the \code{baseline}; e.g., \code{gdd0} when
\code{baseline = 0}), containing values for GDD.
}
}
\description{
Calculate Growing Degree Days above a baseline  \code{baseline}, (GDD) from
daily values of temperature.
}
\examples{
\dontrun{
`\%>\%` <- magrittr::`\%>\%`
data <- tibble::tibble(entity_name = "University of Reading",
                       latitude = 51.44140,
                       longitude = -0.9418,
                       elevation = 61)
smpds::gwr(.ref = "/path/to/reference-tmp.nc",
           .tar = data,
           varid = "tmp") \%>\%
  smpds::pivot_data(varname = "tmp") \%>\%
  smpds::gdd()
}
}
\seealso{
\code{\link{plot_gdd}}

Other utils climate: 
\code{\link{mat}()},
\code{\link{mi}()},
\code{\link{mtco}()},
\code{\link{mtwa}()}
}
\concept{utils climate}
