% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{gwr}
\alias{gwr}
\alias{gwr.character}
\alias{gwr.numeric}
\title{Geographically Weighted Regression}
\source{
This function was adapted from a code developed by Yunke Peng
(\email{yunke.peng@usys.ethz.ch}) - ETH Zürich:
\url{https://github.com/yunkepeng/gwr}
}
\usage{
gwr(.ref, ...)

\method{gwr}{character}(
  .ref,
  ...,
  .tar,
  coordinates = smpds::CRU_coords,
  res = 0.5,
  xy_buffer = 1.5,
  z_buffer = NA,
  cpus = 1,
  bandwidth = 1.06,
  varid = NULL
)

\method{gwr}{numeric}(
  .ref,
  ...,
  .tar,
  coordinates = smpds::CRU_coords,
  res = 0.5,
  xy_buffer = 1.5,
  z_buffer = NA,
  cpus = 1,
  bandwidth = 1.06
)
}
\arguments{
\item{.ref}{Reference data from which the data will be interpolated (see the
details section).}

\item{...}{arguments to be passed to other functions}

\item{.tar}{Table with geographical target data, including: \code{latitude},
\code{longitude} and \code{elevation}.}

\item{coordinates}{Reference data set with columns for \code{latitude},
\code{longitude} and \code{elevation}. Default:
\code{\link{CRU_coords}}.}

\item{res}{Numeric value for the mask resolution. Default: 0.5 degrees.}

\item{xy_buffer}{Numeric value to be used as the boundary for the search
area in the \code{x} and \code{y} axes.
\itemize{
\item \code{latitude} < \code{.tar$latitude + xy_buffer}
\item \code{latitude} > \code{.tar$latitude - xy_buffer}
\item \code{longitude} < \code{.tar$longitude + xy_buffer}
\item \code{longitude} > \code{.tar$longitude - xy_buffer}
}}

\item{z_buffer}{Numeric value to be used as the boundary for the search area
in the \code{z} axis:
\itemize{
\item \code{elevation} <= \code{.tar$elevation * z_buffer}
\item \code{elevation} >= \code{.tar$elevation / z_buffer}
}}

\item{cpus}{Number of CPUs to be used in parallel, default = 1.}

\item{bandwidth}{bandwidth used in the weighting function, possibly
  calculated by \code{gwr.sel}}

\item{varid}{String with the identifier of the main variable inside the
NetCDF file pointed by \code{.ref} (if applicable).}
}
\value{
Table with interpolated values from the \code{.ref} data for each
record/row in \code{.tar}.
}
\description{
Geographically Weighted Regression
}
\details{
The input reference data can be in any of the following formats:
\itemize{
\item \code{Matrix}: this should be a 3-dimensional object with spatial
components (latitude and longitude) and a temporal component for
representing each time step to be used for the extraction of the data.
\item \code{String}: this should point to a valid path on disk where the
reference NetCDF file is stored. Note that the parameter called \code{varid}
should be used to indicate the identifier of the main variable inside the
NetCDF file (e.g., \code{"tmp"},  \code{"pre"},  \code{"cld"}, etc.).
}
}
\examples{
\dontrun{
`\%>\%` <- magrittr::`\%>\%`
data <- tibble::tibble(entity_name = "University of Reading",
                       latitude = 51.44140,
                       longitude = -0.9418,
                       elevation = c(61, 161, 261, 361))
smpds::gwr(.ref = "/path/to/reference-tmp.nc",
           .tar = data,
           varid = "tmp")

ncin <- ncdf4::nc_open("/path/to/reference-tmp.nc")
reference_data <- ncdf4::ncvar_get(ncin, varid)
ncdf4::nc_close(ncin)
reference_data \%>\%
  smpds::gwr(.tar = data)
}
}
\references{
Peng, Y., Bloomfield, K.J. and Prentice, I.C., 2020. A theory of plant
function helps to explain leaf‐trait and productivity responses to elevation.
New Phytologist, 226(5), pp.1274-1284. \doi{10.1111/nph.16447}
}
