% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interpolation.R
\name{gwr}
\alias{gwr}
\title{Geographically Weighted Regression}
\source{
This function was adapted from a code developed by Yunke Peng
(\email{yunke.peng@usys.ethz.ch}) - ETH Zürich:
\url{https://github.com/yunkepeng/gwr}
}
\usage{
gwr(
  .data,
  varid,
  reference = NULL,
  coordinates = smpds::CRU_coords,
  res = 0.5,
  buffer = 1.5,
  cpus = 1,
  bandwidth = 1.06
)
}
\arguments{
\item{.data}{Table with geographical data, including: \code{latitude},
\code{longitude} and \code{elevation}.}

\item{varid}{Main variable name for the NetCDF in \code{reference}}

\item{reference}{Path to the NetCDF file from which \code{varid} will be
loaded.}

\item{coordinates}{Reference data set with columns for \code{latitude},
\code{longitude} and \code{elevation}. Default:
\code{\link{CRU_coords}}.}

\item{res}{Numeric value for the mask resolution. Default: 0.5 degrees.}

\item{buffer}{Numeric value to be used as the boundary for the search area:
\itemize{
\item \code{latitude} < \code{.data$latitude + buffer}
\item \code{latitude} > \code{.data$latitude - buffer}
\item \code{longitude} < \code{.data$longitude + buffer}
\item \code{longitude} > \code{.data$longitude - buffer}
}}

\item{cpus}{Number of CPUs to be used in parallel, default = 1.}

\item{bandwidth}{bandwidth used in the weighting function, possibly
  calculated by \code{gwr.sel}}
}
\value{
Table with interpolated values from \code{varid} for each record/row
in \code{.data}.
}
\description{
Geographically Weighted Regression
}
\examples{
\dontrun{
`\%>\%` <- magrittr::`\%>\%`
data <- tibble::tibble(entity_name = "University of Reading",
                       latitude = 51.44140,
                       longitude = -0.9418,
                       elevation = c(61, 161, 261, 361))
data \%>\%
  smpds::gwr(varid = "tmp",
             reference = "/path/to/reference-tmp.nc")
}
}
\references{
Peng, Y., Bloomfield, K.J. and Prentice, I.C., 2020. A theory of plant
function helps to explain leaf‐trait and productivity responses to elevation.
New Phytologist, 226(5), pp.1274-1284. \doi{10.1111/nph.16447}
}
