% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Herzschuh}
\alias{Herzschuh}
\title{\code{Herzschuh} dataset}
\format{
A data frame (\code{tibble} object) with 595 rows and 17 variables:
\describe{
\item{...}{int ...}
}
}
\source{
https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41467-019-09866-8/MediaObjects/41467_2019_9866_MOESM3_ESM.zip
}
\usage{
Herzschuh
}
\description{
Dataset provided by \insertCite{herzschuh2019position;textual}{smpds}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
