% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IbMPD}
\alias{IbMPD}
\title{\code{IbMPD} dataset}
\format{
A data frame (\code{tibble} object) with 243 rows and 17 variables:
\describe{
\item{...}{int ...}
}
}
\source{
https://researchdata.reading.ac.uk/294
}
\usage{
IbMPD
}
\description{
The Iberian Modern Pollen Dataset (IbMPD).
}
\references{
\insertRef{rdgdr294}{smpds}
}
\keyword{datasets}
