% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IbMPD}
\alias{IbMPD}
\title{\code{IbMPD} dataset}
\format{
A data frame (\code{tibble} object) with 8141 rows and 215 variables:
\describe{
\item{...}{int ...}
}
}
\source{
Harrison, S.P., Shen, Y. and Sweeney, L., 2021. Pollen data and charcoal data of the Iberian Peninsula. University of Reading. Dataset. http://doi.org/10.17864/1947.294
}
\usage{
IbMPD
}
\description{
The Iberian Modern Pollen Dataset (IbMPD).
}
\keyword{datasets}
