% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ItMPD}
\alias{ItMPD}
\title{\code{ItMPD} dataset}
\format{
A data frame (\code{tibble} object) with 79 rows and 196 variables:
\describe{
\item{...}{int ...}
}
}
\usage{
ItMPD
}
\description{
The Italian Modern Pollen Dataset (ItMPD).
}
\references{
\insertRef{finsinger2007modern}{smpds}
}
\keyword{datasets}
