% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-climate.R
\name{mtco}
\alias{mtco}
\alias{mtco.numeric}
\alias{mtco.tbl_df}
\title{Calculate MTCO}
\usage{
mtco(.data, ...)

\method{mtco}{numeric}(.data, ...)

\method{mtco}{tbl_df}(.data, ...)
}
\arguments{
\item{.data}{Object with daily values of temperature. The object can be a:
\itemize{
\item Numeric vector
\item Data frame (\code{tibble} object) with a column called \code{tmp},
with values of daily temperature (a \code{list}).
}}

\item{...}{Optional parameters (not used).}
}
\value{
If the input is a
\itemize{
\item Numeric vector, it returns a single value for MTCO.
\item Data frame (\code{tibble} object), it returns the same data frame
with an additional column called \code{mtco}, containing values for MTCO.
}
}
\description{
Calculate Mean Temperature of the COldest month (MTCO) from daily values
of temperature.
}
