% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NEOTOMA}
\alias{NEOTOMA}
\title{\code{NEOTOMA} dataset}
\format{
A data frame (\code{tibble} object) with 54 rows and 23 variables:
\describe{
\item{...}{int ...}
}
}
\source{
https://www.neotomadb.org
}
\usage{
NEOTOMA
}
\description{
The Neotoma Paleoecology Database subset (NEOTOMA).
}
\references{
\insertRef{williams2018neotoma}{smpds}
}
\keyword{datasets}
