% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SMPDSv1}
\alias{SMPDSv1}
\title{\code{SMPDSv1} dataset}
\format{
A data frame (\code{tibble} object) with 6458 rows and 1569 variables:
\describe{
\item{...}{int ...}
}
}
\source{
Harrison, Sandy  (2019): Modern pollen data for climate reconstructions, version 1 (SMPDS). University of Reading. Dataset. http://dx.doi.org/10.17864/1947.194
}
\usage{
SMPDSv1
}
\description{
SPECIAL Modern Pollen Data for climate reconstructionS, version 1.
}
\keyword{datasets}
